python_parser/requirements.txt
pdfplumber
pandas
openpyxl
openai
python-dotenv
fastapi
uvicorn
python-multipart
